Add SunsetHeader option also to ResourceCollection
